total price to float

+++ db/migrate/20230605214215_create_bookings.rb
@@ -3,7 +3,7 @@ class CreateBookings < ActiveRecord::Migration[7.0]
     create_table :bookings do |t|
       t.date :start_date
       t.date :end_date
-      t.integer :total_price
+      t.float :total_price
       t.references :user, null: false, foreign_key: true
       t.references :venue, null: false, foreign_key: true
